SARS-CoV-2 Antibody Kinetics Model

Create an instance of the SARS-CoV-2 antibody model and fit it to a dataset using Gaussian priors and a given hierarchical model structure.

Methods

Public methods

scova$new()

scova$fit()

scova$simulate_population_trajectories()

scova$population_stationary_points()

scova$simulate_individual_trajectories()

Method new()

Initialise the kinetics model.

Usage

scova$new(
  priors = scova_priors(),
  data = NULL,
  file_path = NULL,
  covariate_formula = ~0,
  preds_sd = 0.25,
  time_type = "relative"
)

Arguments

priors

Object of type 'scova_priors'. Default scova_priors().

data

Optional data table of model inputs. One of data or file must be provided.

file_path

Optional file path to model inputs in CSV format. One of data or file must be provided.

covariate_formula

Formula specifying hierarchical structure of model. Default ~0.

preds_sd

Standard deviation of predictor coefficients. Default 0.25.

time_type

One of 'relative' or 'absolute'. Default 'relative'.

Returns

An epikinetics::sam object.

Method fit()

Fit the model and return CmdStanMCMC fitted model object.

Usage

scova$fit(...)

Arguments

...

Named arguments to the sample() method of CmdStan model. objects: https://mc-stan.org/cmdstanr/reference/model-method-sample.html

Returns

A CmdStanMCMC fitted model object: https://mc-stan.org/cmdstanr/reference/CmdStanMCMC.html

Method simulate_population_trajectories()

Process the model results into a data table of titre values over time.

Usage

scova$simulate_population_trajectories(
  time_type = "relative",
  t_max = 150,
  summarise = TRUE,
  n_draws = 2500
)

Arguments

time_type

One of 'relative' or 'absolute'. Default 'relative'.

t_max

Integer. Maximum number of time points to include.

summarise

Boolean. Default TRUE. If TRUE returns values for 0.025, 0.5 and 0.975 quantiles, if FALSE returns all values.

n_draws

Integer. Maximum number of samples to include. Default 2500.

Returns

A data.table containing titre values at time points. If summarise = TRUE, columns are t, p, k, me, lo, hi, titre_type, and a column for each covariate in the hierarchical model. If summarise = FALSE, columns are t, p, k, t0_pop, tp_pop, ts_pop, m1_pop, m2_pop, m3_pop, beta_t0, beta_tp, beta_ts, beta_m1, beta_m2, beta_m3, mu .draw, titre_type and a column for each covariate in the hierarchical model. lower and upper bounds of titre values; if summarise = FALSE then

Method population_stationary_points()

Process the stan model results into a data.table.

Usage

scova$population_stationary_points(n_draws = 2500)

Arguments

n_draws

Integer. Maximum number of samples to include. Default 2500.

Returns

A data.table of peak and set titre values

Method simulate_individual_trajectories()

Simulate individual trajectories from the model. This is computationally expensive and may take a while to run if n_draws is large.

Usage

scova$simulate_individual_trajectories(
  summarise = TRUE,
  n_draws = 2500,
  time_shift = 0,
  add_variation_params = FALSE
)

Arguments

summarise

Boolean. If TRUE, average the individual trajectories to get lo, me and hi values for the population, disaggregated by titre type. If FALSE return the indidivudal trajectories. Default TRUE.

n_draws

Integer. Maximum number of samples to draw. Default 2500.

time_shift

Integer. Number of days to adjust the exposure date by. Default 0.

add_variation_params

Logical. Default FALSE.

Returns

A data.table. If summarise = TRUE columns are calendar_date, titre_type, me, lo, hi. If summarise = FALSE, columns are stan_id, draw, t, mu, titre_type, exposure_date, calendar_date, time_shift and a column for each covariate in the hierarchical model.